+++ conf/simu.sh
@@ -3,11 +3,11 @@ pip3 install -r ./requirements.txt
 
 # run dask-scheduler and dask-worker
 dask-scheduler --port 8786 &
-dask-worker tcp://localhost:8786 --memory-limit 4GB &
+dask-worker tcp://localhost:8786 --memory-limit 8GB &
 
 # run API server
 ## localhost: gunicorn tide_app:app -k uvicorn.workers.UvicornWorker -b 127.0.0.1:8040 --timeout 120
-gunicorn tide_app:app -w 4 -k uvicorn.workers.UvicornWorker -b 127.0.0.1:8040 --keyfile conf/privkey.pem --certfile conf/fullchain.pem --reload --timeout 120
+gunicorn tide_app:app -w 4 -k uvicorn.workers.UvicornWorker -b 127.0.0.1:8040 --keyfile conf/privkey.pem --certfile conf/fullchain.pem --reload --timeout 180
 
 # kill process
 ps -ef | grep 'tide_app' | grep -v grep | awk '{print $2}' | xargs -r kill -9
@@ -18,3 +18,11 @@ ps -ef | grep -w 'dask-scheduler' | grep -v grep | awk '{print $2}' | xargs -r k
 # pm2 start
 pm2 start ./conf/ecosystem.config.js
 
+## use pipenv
+pipenv update
+
+# update requirements.txt
+pipreqs --force ./
+
+# make local development code updated without reinstalling the package
+pip install -e .